feat : add pulumi code for static web app template
