+++ api/pulumiPrograms/resssource/staticWebApp/README.md
@@ -1,3 +1,3 @@
 # Static Web App Deployment with Pulumi & Azure
 
-This project demonstrates how to deploy a static website (HTML/CSS/JS) to **Azure Storage** using **Pulumi**.
+This project demonstrates how to deploy a static web app using **Pulumi**.
