+++ api/pulumiPrograms/resssource/staticWebSite/Pulumi.yaml
@@ -1,0 +1,16 @@
+name: staticWebSite
+description: "Test a basic template for creating an Azure static website"
+runtime:
+  name: nodejs
+  options:
+    packagemanager: pnpm
+template:
+  displayName: "Create Azure Static Website"
+  description: "A template for creating Azure Static Websites with Pulumi"
+  config:
+    Name:
+      description: "Name of the static website / storage account"
+    location:
+      description: "Azure region"
+      default: "westeurope"
+
