+++ api/pulumiPrograms/resssource/staticWebSite/README.md
@@ -1,0 +1,3 @@
+# Static Web Site Deployment with Pulumi & Azure
+
+This project demonstrates how to deploy a static website (HTML/CSS/JS) to **Azure Storage** using **Pulumi**.
