+++ api/pulumiPrograms/resssource/staticWebSite/index.ts
@@ -1,0 +1,33 @@
+import * as pulumi from "@pulumi/pulumi";
+import * as azure from "@pulumi/azure-native";
+
+const config = new pulumi.Config();
+const projectPrefix = config.require("Name");
+
+const projectPrefixFormated = projectPrefix.toLowerCase()
+
+const resourceGroup = new azure.resources.ResourceGroup(`rg-${projectPrefixFormated}-`, {
+    location: "westeurope",
+});
+
+const storageAccount = new azure.storage.StorageAccount(`st${projectPrefixFormated}`, {
+    resourceGroupName: resourceGroup.name,
+    sku: { name: azure.storage.SkuName.Standard_LRS },
+    kind: azure.storage.Kind.StorageV2,
+    enableHttpsTrafficOnly: true,
+});
+
+const staticWebsite = new azure.storage.StorageAccountStaticWebsite(`stw-${projectPrefixFormated}-`, {
+    accountName: storageAccount.name,
+    resourceGroupName: resourceGroup.name,
+    indexDocument: "index.html",
+});
+
+const indexFile = new azure.storage.Blob("index.html", {
+    resourceGroupName: resourceGroup.name,
+    accountName: storageAccount.name,
+    containerName: "$web",
+    source: new pulumi.asset.FileAsset("www/index.html"),
+    contentType: "text/html",
+});
+export const staticWebsiteUrl = storageAccount.primaryEndpoints.apply(pe => pe.web);
