+++ api/pulumiPrograms/resssource/staticWebSite/package.json
@@ -1,0 +1,12 @@
+{
+    "name": "staticWebSite",
+    "main": "index.ts",
+    "devDependencies": {
+        "@types/node": "^18",
+        "typescript": "^5.0.0"
+    },
+    "dependencies": {
+        "@pulumi/azure-native": "^3.0.0",
+        "@pulumi/pulumi": "^3.0.0"
+    }
+}
